Syntax, ifdef changes. GetCommandSettings() SEND_SETTINGS_MAX changed from 3 to 60.

+++ common/udp.h
@@ -78,7 +78,7 @@ void SetRemoteIP(unsigned long ip);
 
 void SetConfigPort(const std::string& port);
 void SetConfigPort(char* port);
-void SetConfigPort(unsigned int port);
+void SetConfigPort(uint16_t port);
 
 #ifdef D2KCLIENT
 //sends a command packet
